[S05P21C103-134,S05P21C103-137][FE-sjd0051] feat: 프로&스크랩 로그아웃, 반응형 수정

+++ frontend/src/components/Scrap/Scrap.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { Button, Dropdown, DropdownButton } from "react-bootstrap";
 import { GiHamburgerMenu } from 'react-icons/gi';
+import axios from "axios";
 
 import "./css/Scrap.css"
 import "./css/Myscrap.css"
@@ -10,15 +11,19 @@ import { Link } from "react-router-dom";
 class Scrap extends React.Component {
   state = {
     nickname: "",
-    username: ""
+    username: "",
+    profileImg: "",
+    search: "",
   };
 
-  takeUserInfo = () => {
-    if (JSON.parse(sessionStorage.getItem('userInfo'))) {
-      const userInfo = JSON.parse(sessionStorage.getItem('userInfo'))
-      this.setState({nickname: userInfo.nickname})
-      this.setState({username: userInfo.username})
-    }
+  takeUserInfo = () => {    
+    const userInfo = JSON.parse(sessionStorage.getItem('userInfo'))
+    this.setState({
+      nickname: userInfo.nickname,
+      username: userInfo.username,
+      profileImg: userInfo.image
+    })
+
   }
   
   // 검색
@@ -43,14 +48,31 @@ class Scrap extends React.Component {
     }
   }
 
+  // 로그아웃
+  fetchLogout = async (data) => {
+    const url = 'http://localhost:8083/api/auth/logout'
+    await axios.post(url, data)
+    .then(res => {
+      if (res.status === 204) {
+        alert("로그아웃")
+        sessionStorage.clear()
+        window.location.replace("/login");
+      }
+    })
+    .catch(err => {
+      alert("로그아웃에 실패하였습니다😭")
+    })
+  }
+
   // 로그아웃 함수
   onLogout = () => {
-    sessionStorage.clear()
-    window.location.replace("/login");
+    const userInfo = JSON.parse(sessionStorage.getItem('userInfo'))
+    const data = {'userId': userInfo['userId']}
+    this.fetchLogout(data)
   }
 
   render() {
-    const { nickname, username } = this.state;
+    const { nickname, username, profileImg } = this.state;
     const logo = "/image/logo_1.png"
 
     return (
@@ -65,7 +87,8 @@ class Scrap extends React.Component {
             <input className="scrap-search-input" onKeyUp={this.onSearchEnter} autoFocus />
             <button className="scrap-search-btn" onClick={this.onSearch}><h4>🔍</h4></button>
           </div>
-          <div className="scrap-main-profile">
+          {/* 즐겨찾기 상단 내 프로필 */}
+          {/* <div className="scrap-main-profile">
             <div className="scrap-mp-image">
               이미지
             </div>
@@ -76,31 +99,48 @@ class Scrap extends React.Component {
             </div>
           </div>
           <div className="scrap-main-btns">
-            <Link to="/">
-              <Button variant="outline-primary" className="scrap-main-btn" size="sm">알쓸싸잡</Button>{' '}
-            </Link>
+            <Link to="/" className="link"><Button variant="outline-primary" className="scrap-main-btn" size="sm">알쓸싸잡</Button>{' '}</Link>
             <Link to="/edupro">
               <Button variant="outline-primary" className="scrap-main-btn" size="sm">PRO설정</Button>{' '}
             </Link>
             <Button variant="outline-danger" className="scrap-main-btn" size="sm" onClick={this.onLogout}>로그아웃</Button>{' '}
-          </div>
+          </div> */}
+          <div className="scrap-main-profile">
+                <div className="scrap-mp-image">
+                  <img src={ profileImg } alt={ username } className="profile-img" />
+                </div>
+                <div className="scrap-mp-info">
+                  <div className="scrap-mp-nn"><b>{ nickname }</b></div>
+                  <div className="scrap-mp-un">@{ username }</div>
+                </div>
+              </div>
+              <div className="scrap-main-btns">
+                <Link to="/">
+                  <Button variant="outline-primary" className="scrap-main-btn" size="sm">알쓸싸잡</Button>{' '}
+                </Link>
+                <Link to="/edupro">
+                  <Button variant="outline-primary" className="scrap-main-btn" size="sm">PRO설정</Button>{' '}
+                </Link>
+                <Button variant="outline-danger" className="scrap-main-btn" size="sm" onClick={this.onLogout}>로그아웃</Button>{' '}
+              </div>
         </div>
 
+        {/* 햄버거로 돌리기 */}
         <div className="scrap-main-header-responsive">
           <DropdownButton className="hamburger" align="end" variant="secondary" id="dropdown-item-button" title={<GiHamburgerMenu />}>
             <Dropdown.ItemText className="scrap-dd-image">
               <div className="scrap-mp-image">
-                이미지
+                <img src = { profileImg } alt={ username } className="profile-img"/>
               </div>
               <div className="scrap-mr-mp">
                 { nickname }<br />
                 150 point
               </div>
             </Dropdown.ItemText>
-            <Dropdown.ItemText><div className="scrap-line"></div></Dropdown.ItemText>
-            <Dropdown.Item as="button">즐겨찾기</Dropdown.Item>
-            <Dropdown.Item as="button"><Link to="/edupro" className="scrap-link">PRO설정</Link></Dropdown.Item>
-            <Dropdown.Item as="button" onClick={this.onLogout} className="scrap-logout-text">로그아웃</Dropdown.Item>
+            <Dropdown.ItemText><div className="line"></div></Dropdown.ItemText>
+            <Link to="/" className="link"><Dropdown.Item as="button">알쓸싸잡</Dropdown.Item></Link>
+            <Link to="/edupro" className="link"><Dropdown.Item as="button">PRO설정</Dropdown.Item></Link>
+            <Dropdown.Item as="button" onClick={this.onLogout} className="logout-text">로그아웃</Dropdown.Item>
           </DropdownButton> 
           <Link to="/">
             <div className="scrap-main-logo">
